feat(calculator): implement StringParts
